[Add] load_bar/load_tick function for SpreadStrategyEngine
