+++ vnpy/app/spread_trading/engine.py
@@ -5,6 +5,7 @@ from typing import List, Dict, Set, Callable, Any, Type
 from collections import defaultdict
 from copy import copy
 from pathlib import Path
+from datetime import datetime, timedelta
 
 from vnpy.event import EventEngine, Event
 from vnpy.trader.engine import BaseEngine, MainEngine
@@ -17,14 +18,17 @@ from vnpy.trader.object import (
     TickData, ContractData, LogData,
     SubscribeRequest, OrderRequest
 )
-from vnpy.trader.constant import Direction, Offset, OrderType
+from vnpy.trader.constant import (
+    Direction, Offset, OrderType, Interval
+)
 from vnpy.trader.converter import OffsetConverter
 
 from .base import (
     LegData, SpreadData,
     EVENT_SPREAD_DATA, EVENT_SPREAD_POS,
     EVENT_SPREAD_ALGO, EVENT_SPREAD_LOG,
-    EVENT_SPREAD_STRATEGY
+    EVENT_SPREAD_STRATEGY,
+    load_bar_data, load_tick_data
 )
 from .template import SpreadAlgoTemplate, SpreadStrategyTemplate
 from .algo import SpreadTakerAlgo
@@ -1024,3 +1028,25 @@ class SpreadStrategyEngine:
             subject = "价差策略引擎"
 
         self.main_engine.send_email(subject, msg)
+
+    def load_bar(
+        self, spread: SpreadData, days: int, interval: Interval, callback: Callable
+    ):
+        """"""
+        end = datetime.now()
+        start = end - timedelta(days)
+
+        bars = load_bar_data(spread, interval, start, end)
+
+        for bar in bars:
+            callback(bar)
+
+    def load_tick(self, spread: SpreadData, days: int, callback: Callable):
+        """"""
+        end = datetime.now()
+        start = end - timedelta(days)
+
+        ticks = load_tick_data(spread, start, end)
+
+        for tick in ticks:
+            callback(tick)
